♻️ refactor - refatoração de código.

+++ nhyira/src/main/java/API/nhyira/DBA/PersonalRepository.java
@@ -3,5 +3,5 @@ package API.nhyira.DBA;
 import API.nhyira.Model.PersonalModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonalRepository extends JpaRepository<PersonalModel, Long> {
+public interface PersonalRepository extends JpaRepository<PersonalModel, Integer> {
 }
