+++ nhyira/src/main/java/API/nhyira/DBA/UsuarioRepository.java
@@ -1,7 +1,8 @@
 package API.nhyira.DBA;
 
 import API.nhyira.Model.UsuarioModel;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
-}
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {}
